Reuse shared helpers for status tab labels, display formatting and visibility scan
Use data-model's form/field label helpers on the status tab, move
getDisplayDataFromRecordMetadata into formUtilities so the tab stops
importing from the record table, and give currentlyVisibleFields an
includeStaticallyHidden option so statusReport drops its inline scan.
