+++ app/src/utils/formUtilities.ts
@@ -114,6 +114,61 @@ export function formatTimestamp(
   }
 }
 
+/** Placeholder shown when a record value is missing or cannot be rendered. */
+export const MISSING_DATA_PLACEHOLDER = '-';
+
+/**
+ * Converts record metadata field values to displayable strings.
+ *
+ * @param field - The field name to extract from the data
+ * @param data - The data object containing the field
+ * @returns A string representation of the field value, or a fallback value if
+ *          the data is missing or cannot be converted
+ */
+export function getDisplayDataFromRecordMetadata({
+  field,
+  data,
+}: {
+  field: string;
+  data: {[key: string]: any};
+}): string {
+  const fallback = MISSING_DATA_PLACEHOLDER;
+  try {
+    if (!data) return fallback;
+
+    const value = data[field];
+
+    if (value === undefined || value === null) return fallback;
+
+    switch (typeof value) {
+      case 'string':
+        return value.trim() || fallback;
+      case 'number':
+        return Number.isFinite(value) ? value.toString() : fallback;
+      case 'boolean':
+        return value.toString();
+      case 'object':
+        if (Array.isArray(value)) {
+          return value.filter(item => item !== null).join(', ') || fallback;
+        }
+        if (value instanceof Date) {
+          return value.toISOString();
+        }
+        try {
+          const str = JSON.stringify(value);
+          return str === '{}' ? fallback : str;
+        } catch {
+          return fallback;
+        }
+      default:
+        return fallback;
+    }
+  } catch (error) {
+    console.warn(`Error formatting field ${field}:`, error);
+    return fallback;
+  }
+}
+
 /**
  * Converts field names to a more readable format by:
  * 1. Splitting CamelCase into separate words
